#define FIRMWARE_VERSION "1.0.0.2"
bản bao ổn cả font và đả sửa hiện sai vị trí

+++ include/InitStruct.hpp
@@ -18,7 +18,7 @@ struct Product
 };
 
 struct DowntimeItem {
-    int atMeter;
+    float atMeter;
     String objectId;
     String id;
     String jobCode;
